test(pay-card-auth): split platform configuration
chore(pay-card-auth): add ownership and release metadata


chore(pay-card-auth): update lockfile
