+++ .changeset/calm-cards-split.md
@@ -1,0 +1,5 @@
+---
+"@features/flow-pay-card-auth": patch
+---
+
+Split Pay Card authentication configuration and entry points by platform
