+++ features/flow/pay-card-auth/src/components/CardLogin/__tests__/CardLoginView.web.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import { StyleProvider } from "@features/platform-style";
-import { CardLoginView } from "../CardLoginView.web";
+import { CardLoginView } from "../CardLoginView";
 
 const defaultProps: React.ComponentProps<typeof CardLoginView> = {
   title: "Ledger Card",
