+++ features/flow/pay-card-auth/src/components/CardLogin/index.native.tsx
@@ -1,8 +1,10 @@
 import React from "react";
-import { CardLoginView } from "./CardLoginView.native";
+import { CardLoginView } from "./CardLoginView";
 import type { CardLoginProps } from "./types";
 import { useCardLoginViewModel } from "./useCardLoginViewModel";
 
 export function CardLogin(props: CardLoginProps) {
   return <CardLoginView {...useCardLoginViewModel(props)} />;
 }
+
+export type { CardLoginProps, OpenHostedLogin } from "./types";
