+++ features/flow/pay-card-auth/src/index.native.ts
@@ -1,2 +1,2 @@
-export { CardLogin } from "./components/CardLogin/index.native";
-export type { CardLoginProps, OpenHostedLogin } from "./components/CardLogin/types";
+export { CardLogin } from "./components/CardLogin";
+export type { CardLoginProps, OpenHostedLogin } from "./components/CardLogin";
